+++ R/eval-bg.R
@@ -5,6 +5,9 @@
 #' returns immediately.
 #'
 #' @inheritParams r
+#' @param supervise Whether to register the process with a supervisor. If \code{TRUE},
+#'   the supervisor will ensure that the process is killed when the R process
+#'   exits.
 #' @return An `r_process` object, which inherits from [process],
 #'   so all `process` methods can be called on it, and in addition it also
 #'   has a `get_result()` method to collect the result.
@@ -24,11 +27,11 @@ r_bg <- function(func, args = list(), libpath = .libPaths(),
                  repos = c(getOption("repos"),
                    c(CRAN = "https://cloud.r-project.org")),
                  stdout = "|", stderr = "|",
-                 error = c("error", "stack", "debugger"),
+                 error = getOption("callr.error", "error"),
                  cmdargs = c("--no-site-file", "--no-environ", "--slave",
                    "--no-save", "--no-restore"),
                  system_profile = FALSE, user_profile = FALSE,
-                 env = rcmd_safe_env()) {
+                 env = rcmd_safe_env(), supervise = FALSE) {
 
   r_process$new(options = as.list(environment()))
 }