+++ R/options.R
@@ -48,12 +48,13 @@ r_process_options_default <- function() {
     repos = c(getOption("repos"), CRAN = "https://cloud.r-project.org"),
     stdout = "|",
     stderr = "|",
-    error = c("error", "stack", "debugger"),
+    error = getOption("callr.error", "error"),
     cmdargs = c("--no-site-file", "--no-environ", "--slave",
       "--no-save", "--no-restore"),
     system_profile = FALSE,
     user_profile = FALSE,
-    env = character()
+    env = character(),
+    supervise = FALSE
   )
 }
 
@@ -68,7 +69,8 @@ rcmd_process_options_default <- function() {
     system_profile = FALSE,
     user_profile = FALSE,
     env = rcmd_safe_env(),
-    wd = "."
+    wd = ".",
+    supervise = FALSE
   )
 }